requirements.txt
discord.py
yt-dlp
dotenv
PyNaCl
# ffmpeg